+++ packages/in_app_purchases_interface/lib/in_app_purchases_interface.dart
@@ -4,7 +4,6 @@ export 'model/subscription_data.dart';
 export 'model/purchase_state.dart';
 export 'interface/linkfive_interface.dart';
 export 'interface/callback_interface.dart';
-export 'interface/purchase_in_progress_stream_interface.dart';
 export 'interface/purchase_state_stream_interface.dart';
 export 'linkfive/linkfive_environment.dart';
 export 'linkfive/linkfive_logger.dart';