+++ packages/in_app_purchases_interface/lib/interface/purchase_in_progress_stream_interface.dart
@@ -1,3 +1,0 @@
-abstract class PurchaseInProgressStreamInterface {
-  Stream<bool> purchaseInProgressStream();
-}